Classi BL e Testing
-Sistemati i costruttori allineando il tipo di ID
-GetAllOttoni su ClsOttoneBL e testing del metodo su FrmHome

+++ NegozioStrumentiMusicali_Cappelloni-DiBernardo/DMO/ClsStrumentoACorda.cs
@@ -251,7 +251,7 @@ namespace NegozioStrumentiMusicali
 	        
 	    }
         //Costruttore ereditato dalla classe madre
-        public ClsStrumentoACorda(int id) : base(id)
+        public ClsStrumentoACorda(long id) : base(id)
         {
 
         }
